[Refactor] change directory structure
