+++ srcs/game/game_setup.c
