+++ srcs/mino/clear_mino.c
