+++ srcs/mino/generate_mino.c
