+++ srcs/mino/mino.c
